% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_accn.R
\name{Accn}
\alias{Accn}
\alias{Accn2}
\alias{AccnStack}
\alias{AccnMultiStack}
\title{Accelerometer ("Accn") constructor classes for MoveRakeR}
\usage{
Accn(data)

Accn2(data)

AccnStack(data)

AccnMultiStack(data)
}
\arguments{
\item{data}{An input base::data.frame for an accelerometer dataset that is to be
coerced to the designated class.}
}
\value{
For tracking data, returns a \code{Accn}, \code{AccnStack} or
\code{AccnMultTrackStack}
}
\description{
Functions to append object classes of \code{Accn}, \code{AccnStack} and \code{AccnMultiStack},
obtained respectively using their respective namesake functions respectively. this is a specific
class for the UvA accelerometer sensor data.
}
\details{
Three special object classes are returned from initial the read in function
\code{\link{read_accn_UvA}}. The class \code{Accn} is typically representative
of an individual animal's time-series, but an 'rbinded' \code{Accn}
object of several animals, i.e. multiple TagIDs, will also be accepted and processed for each TagID,
thus \code{Accn} objects represent a single \code{data.frame} class;
\code{AccnStack} and \code{AccnMultiStack} objects are nested \code{list} formats, and (as currently implemented)
\emph{must be treated as one individual per data.frame}. These stacked listed methods
are intended for efficient application of routines for groups of animals. Objects can be coerced into
\code{Accn}, \code{AccnStack} and \code{AccnMultiStack} classes but they require the following
minimum column names: \strong{TagID}, \strong{DateTime} and relevant xyz metrics
for a minimum 3D accelerometer: \strong{x_acceleration}, \strong{y_acceleration}, \strong{z_acceleration}.
Further checking of whether objects belong to special classes are provided by:
\code{link{is_Accn}},
\code{link{is_AccnStack}} and \code{\link{is_AccnMultiStack}}, and coercing of two
\code{AccnStack} objects is provided by \code{\link{accn_multi_stack}}. See \code{?Track} for details
of \code{Track} family classes.
}
\seealso{
\link{Track}, \link{Pr}, \link{accn_multi_stack}, \link{is_Accn}, \link{is_AccnStack}, \link{is_AccnMultiStack}
}
