% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_accn.R
\name{accn_multi_stack}
\alias{accn_multi_stack}
\title{accn_multi_stack}
\usage{
accn_multi_stack(data = NULL)
}
\arguments{
\item{data}{Object to be assesed.}
}
\value{
Return new \code{data} coerced to a \code{\link{AccnMultiStack}} class.
}
\description{
Coerce \code{AccnStack} class objects into a \code{AccnMultiStack} class object.
}
