% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_track.R
\name{clean_track}
\alias{clean_track}
\alias{clean_GPS}
\alias{get_attributes}
\alias{give_attributes}
\title{Cleaning the tracking data}
\usage{
clean_track(
  data,
  min_fix = 0,
  min_fix_when = c("after_NA", "after_all"),
  NA_latlong = TRUE,
  sat_check = TRUE,
  drop_sats = 3,
  sat_col = "satellites_used",
  pdop_check = FALSE,
  pdop_val = 7,
  pdop_col = "gps.pdop",
  hdop_check = FALSE,
  hdop_val = 7,
  hdop_col = "gps.hdop",
  flt_switch = FALSE,
  custom_filt = FALSE,
  hasgaps = FALSE,
  GAP = 28800,
  tol = 0.2,
  drop_single_gap = FALSE,
  duplicates = TRUE,
  dup_method = c("random", "best_quality", "first", "last", "mean"),
  tagspans = TRUE,
  measure = c("spatial", "temporal"),
  speedFilt = TRUE,
  sp_thres = 50,
  turnFilt = NULL,
  turnFilt_dir = c("less_than", "greater_than"),
  annotate = FALSE,
  attrib = TRUE,
  reproject = FALSE,
  p4s = 3035,
  verbose = TRUE,
  detailed_verbose = FALSE,
  ...
)

clean_GPS(
  data,
  min_fix = 0,
  min_fix_when = c("after_NA", "after_all"),
  NA_latlong = TRUE,
  sat_check = TRUE,
  drop_sats = 3,
  sat_col = "satellites_used",
  pdop_check = FALSE,
  pdop_val = 7,
  pdop_col = "gps.pdop",
  hdop_check = FALSE,
  hdop_val = 7,
  hdop_col = "gps.hdop",
  flt_switch = FALSE,
  custom_filt = FALSE,
  hasgaps = FALSE,
  GAP = 28800,
  tol = 0.2,
  drop_single_gap = FALSE,
  duplicates = TRUE,
  dup_method = c("random", "best_quality", "first", "last", "mean"),
  tagspans = TRUE,
  measure = c("spatial", "temporal"),
  speedFilt = TRUE,
  sp_thres = 50,
  turnFilt = NULL,
  turnFilt_dir = c("less_than", "greater_than"),
  annotate = FALSE,
  attrib = TRUE,
  reproject = FALSE,
  p4s = 3035,
  verbose = TRUE,
  detailed_verbose = FALSE,
  ...
)

get_attributes(data)

give_attributes(data, attr_list)
}
\arguments{
\item{data}{Input data object, with required columns: TagID, DateTime, longitude, latitude.}

\item{min_fix}{Minimum number of fixes to identify as valid or retain per animal. The default is 0,
i.e. all fixes are retained. This parameter is used in conjunction with min_fix_when to determine where in the process this is applied.}

\item{min_fix_when}{Where to apply the minimum number of fixes per animal (min_fix) criterion. Two options are
available: "after_NA" carries out this assessment after initial NA rows and NA lat longs have been removed, and "after_all"
makes the calculation after all other filters have been applied.}

\item{NA_latlong}{logical condition (default TRUE), whether to annotate or remove NA latitude, longitude records,
although check if this is compatible with further analyses downstream.}

\item{sat_check}{Logical whether to filter data for satellite count, defaults to TRUE, and if TRUE,
the data needs a column called matching that supplied under 'sat_col'.}

\item{drop_sats}{Numeric integer value for the minimum number of GPS satellites to drop from the database.
Defaults to 3, meaning only 4 and above are flagged as suitable, or retained in the data.}

\item{sat_col}{The column to search for satellite counts, defaults to "satellites_used".}

\item{pdop_check}{Logical (default FALSE), whether to activate the pdop filter.}

\item{pdop_val}{Default is numeric 7, if pdop_check = TRUE, the function will look for a column named under the 'pdop_col' argument,
and keep only data less than that value.  NAs are present these are retained in the outputs.}

\item{pdop_col}{Character variable (default "gps.pdop") directing the column to be used for the pdop filter.}

\item{hdop_check}{Logical (default FALSE), whether to activate the hdop filter.}

\item{hdop_val}{Default is numeric 7, if hdop_check = TRUE, the function will look for a column named under the 'hdop_col' argument,
and keep only data less than that value.  NAs are present these are retained in the outputs.}

\item{hdop_col}{Character variable (default "gps.hdop") directing the column to be used for the hdop filter.}

\item{flt_switch}{logical argument, whether to annotate/exclude
MoveBank data erroneous flt_switch values. Warning though this can remove fixes erroneously if not careful
so it is suggested if not sure to look at Movebank directly and do this separately. This is set to
FALSE by default i.e. no flt_switch filtering done.}

\item{custom_filt}{Logical (default FALSE) whether to use any further custom columns to filter/annotate by.
This process if fully flexible for numeric or character/factor variables and uses the \code{file_err()} function
hence to use this option, the necessary arguments must be specified for that function. These are the names
of the variables to use via \strong{nms}, the lower bound via \strong{vals_l}, the upper bound via
\code{vals_u}, and optionally the \strong{inclusive} argument, for the upper and lower bound condition. These arguments
are supplied to \code{filt_err} via the dots (...) argument within \code{clean_track}. Note also,
if custom_filt is TRUE, checks are first made for use of prior built-in filters via the number of satellite, pdop,
hdop and flt_switch checks, and if any duplicate column names are found in the 'nms' argument those having been
used already, these are omitted from the custom filter.}

\item{hasgaps}{A logical vector whether the data being fed in already has gap sections defined from another process.
If FALSE, \code{clean_track} will assign its own gap sections based on the values in GAP and tol.}

\item{GAP}{Numeric integer in seconds. The function can label sections of assumed continuous GPS recording
for the individual animal, using the \code{\link{gap_section}} function.
The argument \code{GAP} here is the cut-off i.e. temporal hiatus between two
consecutive GPS fixes, which defaults to 28800 seconds (eight hours), although note, the tolerance parameter
will also be incorporated (see tol argument).
This is a subjective choice for the user, and will depend also on GPS sampling
resolution. If a precise GAP is needed with no tolerance, set tol to 0. Units are in seconds.}

\item{tol}{A tolerance is also applied to the GAP value, useful for example to give lee-way around
the gap criteria relating to imprecision of sampling rate in GPS deviating from the initial rate specified.
By default this is set as 0.2, i.e. to give a true value used in delineation as: GAP +/- (GAP*0.2) s}

\item{drop_single_gap}{logical defaults to FALSE, but user can specify TRUE to annotate/retain single gap sections
that by default get removed from the speed filter as isolated fixes that are at too coarser rate.
as determined by the \code{GAP} argument. This replaces the previous 'drop' argument in MoveRakeR <1.0.2.}

\item{duplicates}{logical defaulting to TRUE to call \code{duplicate_track()} to address data duplicates
Also note that this will flag/remove any duplicated rows if you are working with trips that 'come-and-go'
from a central place; see function: \code{add_cag_trips}.}

\item{dup_method}{Corresponding to \code{duplicate_track()} argument. Global conflict–resolution
dup_method used for all duplication steps unless overridden by a further method-specific duplication call (see examples).
dup_method is one of \code{"best_quality"}, \code{"first"}, \code{"last"}, \code{"random"}, or step-specific dup_methods
where applicable. Defaults to "random" if nothing is provided; further formal arguments for \code{duplicate_track()}
can be passed via the '...' of \code{clean_track()}.}

\item{tagspans}{logical defaulting to TRUE to call \code{tag_spans()} whether to consider a priori reasons why
data may be erroneous in time or space; further formal arguments for \code{duplicate_track()}
can be passed via the '...' of \code{tag_spans()}.}

\item{measure}{corresponds to the equivalent argument in \code{tag_spans()} one of c("spatial", "temporal") fro
consideration of temporal and/or spatial extent flagging or resolving.}

\item{sp_thres}{As supplied to \code{speed_filt()}. A numeric value for the minimum threshold (m/s) to be used in the speed filter, which requires
inspection and will vary by species. This will depend on the sampling resolution
of your data. The function then uses a while loop to remove all points not
meeting the threshold level until all offending fixes are removed.}

\item{turnFilt}{An optional filter, default set to NULL (no filter applied) to annotate or filter using \code{turn_filt()};
turning angles greater or lesser than a certain "turn_filt" turning angle provided are flagged/filtered,
but beware for GPS fixes for a stationary animal, as turning angles represent GPS error that can be erratic, and so this may
need further behavioural consideration. If supplied, turn_filt should be a single value with a maximum of pi, that
represents the modulus of the radian turning angle; note also that if turn_filt_dir = "less_than" and turn_filt = pi, then
all fixes would be removed; A value of  turn_filt = pi/4 and turn_filt_dir = "less_than" would retain all fixes less_than turn_filt.}

\item{turnFilt_dir}{As supplied to \code{turn_filt()}, a direction indicating greater or lesser than the turn_filt value to annotate/retain values.}

\item{annotate}{Logical defaulting to FALSE, in which case data will be \strong{filtered}, however if TRUE, data
will be \code{annotated} instead retaining all rows, and different columns will be added for each of the \code{clean_track}
manipulation steps representing rows flagged under each condition (see 'Value' section).}

\item{attrib}{Logical set to FALSE if detailed attributes should be attached
for numbers of fixes removed in various stages from \code{clean_track}.
These can be accessed afterwards using: attr(data, "data_summary").}

\item{reproject}{As above, a logical value whether or not to reproject lat-long WGS84
coordinates to the new desired projection as specified via, \code{p4s}.}

\item{p4s}{The coordinate reference system projection, should it be needed, defaults to NW Europe epsg:3035.}

\item{verbose}{Logical argument if run-time messages are wanted.}

\item{detailed_verbose}{Logical argument if the very detailed process of the speed filtering
part of the function is required. Defaults to FALSE.}

\item{...}{Additional arguments that can be passed on to function, which at the moment is limited mainly to
\code{speed_filt()}, where additional arguments: "hybrid", "fall_back", "include_centre", "method",
"fun", "fun_args" and "points_each_side" can be passed on lexically feeding through to lower functions.}

\item{speed_filt}{logical defaulting to TRUE as to whether to use the speed filter and call \code{speed_filt()}.
\code{clean_track} calls the \code{speed_filt} function to do this, and in turn further arguments can be fed
through to \code{speed_filt} from \code{clean_track} via the dots (...) argument; these include the option to
use further root mean square or smoothing options such as through the \strong{method} argument of \code{speed_filt}
and whether to use both raw trajectory speed and the 'smoothed' version via the \strong{hybrid} argument.}
}
\value{
The function returns a tibble format with a \code{Track} class.
Resultant output depends on the argument \code{annotate}; if TRUE, then data are filtered for
the required processing steps selected in \code{clean_track}, with additional columns in the objects if not already included:
including: \strong{dt}, time difference (secs) between consecutive GPS fixes,
\strong{gap}, a binary (1,0) for fixes where a gap was identified and \strong{gapsec}
as a numeric integer of sequentially-labelled gapsection strings of telemetry fixes. If the \code{annotate} argument is
TRUE, then various columns can be added flagging as a 1 or 0 (some with NA conditions) whether a fix
was labelled as 'incorrect' through the thresholds supplied. These columns are named accordingly for conditions:
\itemize{
\item tid_rm (minimum no. animals);
\item NA_latlong_rm (lat/long point NA);
\item all_NA_rm (NA rows);
\item sat_NA (NA of rows without satellites);
\item sat_rm (number of satellites);
\item flt_rm (flt_switch);
\item gps_pdop_rm (pdop);
\item gps_hdop_rm (hdop);
\item singlegap_rm ('isolated' gapsections);
\item dup_rm (duplicate DateTimes);
\item speed_rm (speed filter);
\item angle_rm (angle filter);
\item if using the 'custom_filt' argument, any additional columns used will be flagged with a suffix "_rm";
\item if using the duplication method, then a further five columns ar possible relating to specific
issues identified from the duplication process: dup_parallel_rm (parallel issue), dup_time_rm (same DateTime, different coordinates),
dup_coords_rm (same coordinates, different DateTime), dup_time_coords_rm (both DateTime and coords the same)
and all_dup_rm (all row duplicates);
\item if using tagspans a further six columns are possible: future_date_rm (future date found), temporal_date_rm (dates exceed supplied range),
minx_rm (< minx), miny_rm (< miny), maxx_rm (> maxx), maxy_rm (> maxy)
and then depending on conditions returned:
\item 'combined_rm' is the result of the combined flags (1,0,NA) across all "_rm" columns.
}
}
\description{
\code{clean_track} is provided to allow some basic cleaning of the raw tracking data to be
carried out, i.e. after having read in the raw data using
\code{\link{read_track_UvA()}} and/or \code{\link{read_track_MB()}}. \code{clean_track()} is used
to address a variety of issues with tracking data such as (1) sampling issues like including data duplicates,
missing data, errant spatio-temporal data spans, numbers of fixes per animal, and gaps between valid strings
of points; as well as (2) further issues relating to measurement error and unrealistic movement trajectories.
}
\details{
\strong{Introduction}

A wrapper function \code{clean_track()} is provided that cycles through several main steps to either annotate
potential 'bad' data rows, or filter such rows, thus 'cleaning' the GPS data. As an initial upfront disclaimer,
\code{clean_track()} calls many other \code{MoveRakeR} functions and each of these can be also used in
their own right outside of \code{clean_track()}. Therefore, the \code{clean_track()} wrapper may be more useful once you are sure the process
of annotating or filtering data is "locked in" and you have done all the necessary checks. \code{clean_track()}
may therefore be more useful as a workflow-tidier, or if you are just certain of how to clean the data then that's totally fine.
As always, this is a gradient from fully open code (i.e. outside of \code{MoveRakeR}) to using
the \code{MoveRakeR}-family of functions through to then using the fully-wrapped up process of \code{clean_track()}.

There many reasons why (here focusing on GPS) animal-borne telemetry locations may have error, such as
internal firmware settings of tags e.g. connection parameters with satellites, sampling rate
of tags, number of satellites, satellites position in the sky, along with external factors
such as temperature, humidity etc. Here, we assume the user is mainly interested in 2D space,
so we retain all vertical information in columns \code{altitude} and \code{altitude_agl}. Further
filters for such z-dimension columns may be needed by the user.

Here, we carry out basic cleaning to annotate or filter fixes that may be identified less reliable.
The steps within \code{clean_track} are often more simple trivial data manipulations, that can be accomplished in your own way
However, many of these steps can quickly get more complicated than originally intended. \code{MoveRakeR} therefore
is there to help if that is the case. Annotation is also at the heart of \code{clean_track}
as it is expected that other functions from neighbouring R packages may come in handy to assess
alongside this process.

\strong{The general process}

Within \code{clean_track}, there are four main 'groups' of cleaning defined here:
\enumerate{
\item Duplicate and missing data Location and spatio-temporal spans of data;
\item Sample size and minimum numbers of animals and fixes per animal, and data gaps between points;
\item Error covariates, such a number of satellites, dilution of precision or other accuracy measures;
\item Further location filters based on speed/ and/or turning angle (building on 2. above);
}

Under each one of these banners, \code{clean_track} handles (in certain order):
Location and timestamp issues
\itemize{
\item Missing location - any missing latitude/longitude information (NAs) are first identified and flagged/filtered;
\item Missing data - Label/drop rows where all columns in the data are NA.
\item Duplicate date issues - using the function \code{duplicate_track()} duplicate rows are either annotated or resolved.
\item Data spans are assessed using the \code{tag_spans()} function, which can assess errant future dates, specific dates
supplied by the user, and a priori spatial extent expectations.
}

Sample sizes and data gaps
\itemize{
\item Minimum number of GPS fixes per animal;
\item Gap section labeling using the \code{gap_section} function (data annotation only), flagging user-defined gaps in the data and in turn:
\item Identification of or removal of 'orphaned' GPS fixes that have only one GPS fix per gap section (potentially problematic for further analyses);
}

Error based on covariates
\itemize{
\item A value for the minimum number of satellites can be provided below which data are flagged/filtered;
\item Satellite position, via the GPS 'hdop' and 'pdop' values if present in the data, with a threshold value above which data are flagged/filtered;
\item A Movebank-specific condition based on rows flagged in the database as 'flt_switch' problem rows can be flagged/dropped;
\item Bespoke accuracy filters via available columns in the data, via the \code{filt_err()} function.
}

Speed and turning angle filters
\itemize{
\item Trajectory (ground) speed filter, that calls both \code{traj_speed()} to calculate speeds within identified gap sections, and \code{speed_filt()} to
carry out labeling/filtering of points beyond unrealistic perceived animal travel speeds. Root-mean square or
other smoothing operations can be fed through to \code{traj_speed()} via \code{clean_track()};
\item A turning angle filter, using function \code{turn_filt()} to annotate/filter fixes more ore less than a modulus(radians) turning angle between successive fixes.
The turning filter also depends on \code{adehabitatLT::as.ltraj} for efficiently computing turning angles and
should be used with a high degree of caution given GPS location of stopped animal movement reflects GPS noise.
}

\strong{Further details}

As a bove, the defaults of \code{clean_track()} should ideally not be relied upon 'out-of-the-box'. Selection of parameters should be carefully
considered and scutinised (i.e. raking) through the use of the \code{ShinyRakeR} app.

Contexts will likely vary greatly and will be highly species- specific.
Ground speeds will of course vary considerably by species - so it may be best to view
the distribution, with the help of function \code{rake_outlie}, to identify outliers through simple numeric statistics.
Similarly for all perceived biases in the data, viewing of the distribution of the variables that could be driving
such variation may be useful to view using \code{rake_outlie}, such as the pdop and hdop thresholds, if they are deemed important.

The raw trajectory speed filter (default) is sensitive to single 'bad' spikes. However, the speed filtering process can be
expanded to consider instances where several fixes in a row may be slightly 'wrong' but falling below the extreme outlying cases
identified via raw trajectory speed. It may therefore be necceassary to ask a slightly different question:
"is there sustained implausible movement in the' neighbourhood for this fix?".
Further arguments can therefore be supplied to \code{clean_track} to address this, by using
a root mean square or smoothing of trajectory speed, which alongside the raw trajectory speed amounts to a "hybrid" approach
to isolate individual bad spikes as well as local context-dependent neighbourhood error. See also \code{trip::speedfilter}
for a similar approach.

Trajectory speed filters naturally have drawbacks as they are tied intrinsically to the sampling rate of the tag. Information
on instantaneous speed may be preferable, although not provided in \code{clean_track} as a filtering/annotating option.
Coarser GPS sampling rates may make it harder to discern potentially erroneous fixes given a wider time elapsed over what the animal could have been doing and
at very fine spatio-temporal scales, i.e. with fast sampling less than 60 seconds, a small change in GPS position has a greater influence on trajectory speed
than for coarser sampling schedules. It may therefore be needed to investigate the error potential in traj_speed calculations at these finer scale
and the potential sensitivity of the data to single value decisions. This relationship can be examined further
using the \code{tspeed_jit} function. Note this may be resolved partly by considering a root mean square/smoothing
operation within the speed filter (see examples), increasing robustness over different sampling regimes.

The \code{clean_track} workflow in \code{MoveRakeR} is not the first of its kind to have dealt with such similar issues.
There are a number of packages that have functions that may be of use to consider as annotation of the data alongside
\code{clean_track}, such as function \code{trip::sda} and \code{trip::speedfilter} that have methods for speed/turning angle point assessment;
the function in s \code{MoveRakeR::clean_track} differ for trajectory speed in using a while loop to sequentially assess potentially
erroneous fixes that may be above the maximum speed for the species, as strings of erroneous points can occur, thus being displaced
from where they likely should have been; hence the  \code{MoveRakeR::speed_filt} by design also considers a displacement effect indirectly.

Ultimately these filtering and annotating steps are a little subjective, and depends on the purpose of the analysis.
It is not the intention of \code{MoveRakeR} however to re-invent all existing methods, but instead add to the tools available that we can assess
biases, and so the user is also directed to other packages such as \code{trip}, \code{traipse}, \code{move/move2} and \code{AMT}.

\emph{NOTE as of v 1.1.3.90000}
\code{clean_track()} used to be called \code{clean_GPS()} in previous \code{MoveRakeR} releases.
The function is the same, this is just a name change, and users can still
call \code{clean_GPS()}, but please consider that older name deprecated.

\strong{Wider discussion on data cleaning/raking}

There are different coding approaches to tackle cleaning of GPS data. It is fully appreciated users may have their own
approaches to do this. If these process are done in 'open code', following the \code{ExMove} workflow (Langley et al. 2024)
this is useful starting point. The user may then hit repeatable issues we all face, for which the functions in
\code{MoveRakeR} seek to simplify and capture to avoid verbose coding. These functions are then
brought together in \code{clean_track()} as an overall wrapper. In between these steps, there is also the
defined 'raking' of data to explore what happens if you choose option X, or parameter Y, i.e. how much data
is dropped, do these exclude certain individuals or time periods etc. Data cleaning can be seen on a continuum of approaches ranging
from the fully open-code approaches (that may result in repeated coding of the same sorts of issues), to a fully wrapped up
approach of \code{MoveRakeR::clean_track} as a single function for which use 'out-of-the-box') is risky without first delving into the
error present in the data. It is also recognised that \code{MoveRakeR} is coded around examples on real GPS data for which we
have detected. However, there may be unknown issues we have not yet come across, that a user may switch to open coding of
data cleaning. Do feel free to get in touch if there are any steps that could be included.
}
\examples{

indata <- yourdata # data.frame with a minimum of columns named TagID, DateTime, longitude, latitude
data <- Track(indata) # optional to convert to in-house class (most functions will still run on non track data but generic S3 methods will not be available for plot and summary)
data <- data \%>\% clean_track() # accepting the defaults of the clean_track function

############################### DISCLAIMER ####################################
# clean_track() is a wrapper for several other MoveRakeR functions
# It is advisable to carry out piecemeal checks of data before using clean_track() but thereafter if
# the filtering or annotating process is "locked in" and you are sure of the steps used, then
# clean_track() could be useful as shorthand to help clean up workflows

# --------------------------------------------------------------------------- #
# Example of basic use
# --------------------------------------------------------------------------- #
# as an annotation tool:
data_an <- clean_track(data,
 min_fix = 0, min_fix_when = "after_NA",                       # keep birds with 5 fixes or more, after NA annotation applied
 NA_latlong = TRUE,                                            # drop NA lat-longs
 sat_col = "satellites_used", sat_check = TRUE, drop_sats = 3, # drop fixes with 3 satellites
 pdop_check = FALSE, pdop_val = 7, pdop_col = "gps.pdop",      # drop any rows above a value of 7 IF pdop_check TRUE
 hdop_check = FALSE, hdop_val = 7, hdop_col = "gps.hdop",      # drop any rows above a value of 7 IF hdop_check TRUE
 flt_switch = FALSE,                                           # flt_switch custom argument, ignored here
 custom_filt = FALSE,                                          # should any custom filtering be done, no here but nms, vals_l, vals_u would then be needed (see below)
 hasgaps = FALSE, GAP = 28800, tol = 0.2,                      # no gaps previously defined, if TRUE use GAP and tol args to redefine
 drop_single_gap = FALSE,                                      # keep all gap sections identified with 1 fix (note these are dropped anyway in speed filter if not annotating)
 duplicates = TRUE, dup_method = "random",                     # activate duplicate filter, defaulting to a random pick in dup_method
 tagspans = TRUE, measure = c("spatial", "temporal"),          # activate the tagspans data checks (default spatial and temporal)
 speedFilt = TRUE, sp_thres = 50,                              # activate basic speed filtering using only raw GPS traj speed - see further options for smoothed and RMS consideration
 turnFilt = NULL, turnFilt_dir = "less_than",                  # no use of turn filter under default
 annotate = TRUE,                                              # default is usually FALSE but TRUE annotates rather than filters
 attrib = TRUE, reproject = FALSE, p4s = 3035,                 # attributes and p4s conditions
 verbose = TRUE, detailed_verbose = FALSE)                     # verbosity

# If you are NOT annotating, you can check where fixes were dropped via:
attr(data_an,"data_summary")
# otherwise you can assess this using the '_rm' columns in the annotated dataset:
data_an$combined_rm # the combined annotation column

# --------------------------------------------------------------------------- #
# Further use of custom filtering via the dots(...) argument of clean_track()
# --------------------------------------------------------------------------- #
# The below call activates 'custom_filt' and looks for columns speed_accuracy, flag ("wrong" or "correct"), and satellites_used
# but... notices how sat_check is already TRUE and satellites_used is specified there so would be ignored by
# the custom filterer. Note also the vals_l and vals_u notation; here anything >= 11 speed_accuracy would be filtered as 'incorrect' (annotate = FALSE = destructive filtering)
# If sat_check was FALSE, then anything less than or equal to 3 satellites would be dropped via the vals_l argument.
# If the flag column is "wrong", then these would be dropped. vals_l and vals_u are best supplied as lists if multiple
# columns are to be used.

data$flag <- "0" # IF YOU DO NOT HAVE A FLAG COLUMN (used in this example), otherwise filt_err() will fail.

cust_filt = clean_track(data,
                sat_check = TRUE, drop_sats=3, sat_col = "satellites_used",
                duplicates = TRUE,  tagspans = TRUE,
                annotate = FALSE, # <--- i.e. dropping data that are flagged as errant
                pdop_check = TRUE, pdop_val = 9, pdop_col = "gps.pdop",
                hdop_check = TRUE, hdop_val = 9, hdop_col = "gps.hdop",
                custom_filt = TRUE,
                nms = c("speed_accuracy", "flag", "satellites_used"),
                vals_l = list(NULL,NULL,3),
                vals_u = list(11,"wrong",NULL),
                drop_single_gap = TRUE, hasgaps = TRUE, GAP = 15000, tol = 0.2,
                speedFilt = TRUE, sp_thres = 50)

# Custom functions can be fed through to any column you want - see filt_err() for more details
# Here, a potential custom function use on the 'flt_switch' in the data
flag_flt_switch <- function(x) {
 flag <- as.integer(!(is.na(x) | x == 0))
 return(flag)
}

cust_filt2 = clean_track(data,
                sat_check = TRUE, drop_sats=3, sat_col = "satellites_used",
                duplicates = TRUE,  tagspans = TRUE,
                annotate = FALSE,
                pdop_check = TRUE, pdop_val = 9, pdop_col = "gps.pdop",
                hdop_check = TRUE, hdop_val = 9, hdop_col = "gps.hdop",
                custom_filt = TRUE,
                nms = c("speed_accuracy", "flag", "satellites_used", "flt_switch"), # <-------- "flt_switch" now added to c() of nms
                vals_l = list(NULL,NULL,4,NULL),                                    # <-------- extra NULL lower condition
                vals_u = list(11,"wrong",NULL,flag_flt_switch),                     # <-------- 'flag_flt_switch' function specified
                drop_single_gap = TRUE, hasgaps = TRUE, GAP = 15000, tol = 0.2,
                speedFilt = TRUE, sp_thres = 50)

# --------------------------------------------------------------------------- #
# Speed filter further option via the dots(...) argument of clean_track()
# --------------------------------------------------------------------------- #
# Similar to the use of the 'custom_filt' argument above, the speed filterer in clean_track can also accept
# further arguments via '...' that are fed through to speed_filt(). In the example below we annotate the data
# and use the speed filterer but also specify that we want to consider a hybrid approach using raw
# trajectory speed AND the root mean square "mcconnell" method; this uses two fixes either side back and forward (excluding the current fix)
# The RMS/smooth alongside the raw traj speed is useful to also search for potentially erroneous
# sustained implausible fixes around the current fix indicating local neighbourhood inconsistency
# as well as 'spikes' in the data from the raw trajectory filter.

cust_speed = clean_track(data,
                sat_check = TRUE, drop_sats=3, sat_col = "satellites_used",
                duplicates = TRUE,  tagspans = TRUE,
                annotate = FALSE,
                pdop_check = TRUE, pdop_val = 7, pdop_col = "gps.pdop",
                hdop_check = TRUE, hdop_val = 7, hdop_col = "gps.hdop",
                speedFilt = TRUE, sp_thres = 50, method = "mcconnell", hybrid = TRUE) # <------ supplied to speed_filt() via '...'

# --------------------------------------------------------------------------- #
# Duplicates
# --------------------------------------------------------------------------- #
# This calls the duplicates_track() function
# However the only global argument for this process in clean_track() is
# dup_method = c("random", "best_quality", "first", "last", "mean")
# But you can make use of duplicates_track() by specifying any argument that function needs by just
# specifying them in the clean_track() function which will be fed through as above with other process via the "..."

# You can use the single functions anyway but for example specifically using
# "best quality" as the dup_method BUT then also adding which columns through further variables:

data_dups = clean_track(data, duplicates = TRUE, annotate = TRUE,
  dup_method = "best_quality",
  quality_vars = c("satellites_used", "accuracy", "pdop", "hdop"),
  quality_directions = c(1,1,-1,-1)
)

# or even more bespoke, using a specific method that is allowed for a particular dup_method
# these are passed through the 'dots' of the duplicate_track() essentially double-dotted from clean_track():
data_dups = clean_track(data, duplicates = TRUE,
  dup_method_same_time = "centroid",
  annotate = FALSE,
  verbose = TRUE # default
)
# which should show "Using dup_method: centroid" in the Duplicate data checker if verbose is TRUE

# --------------------------------------------------------------------------- #
# Tag data spans
# --------------------------------------------------------------------------- #
# Tag data span checks are computed using the tag_spans() function
# This also makes use of a single global argument in clean_track() but as above
# clean_track() accepts the necessary further functions within tag_spans():

data_tagspans <- clean_track(data,
    tagspans = TRUE,
    measure = c("spatial", "temporal"), # <- general summaries in tag_spans() cannot be used for clean_track()
    xra = c(-10,5), yra = c(25,65),
    dates = c("2014-01-01 00:00:01", "2021-12-31 23:59:59"),
    annotate = FALSE
  )

# --------------------------------------------------------------------------- #
# Outside of the clean_track() function
# --------------------------------------------------------------------------- #
# As noted above, clean_track() is a wrapper for several other MoveRakeR functions
# It is advisable to carry out piecemeal checks of data before using clean_track()
# If the user wants maximal control over the process and in particular the ORDER in
# which different cleaning steps are applied then best to do outside of clean_track()

# Example, using the speed filter, outside of clean_track() with annotation:

# define gapsections
data <- gap_section(data,GAP=28800, tol = 0.2)

# use speed filter, with the above gaps, annotating the data
data_speed_filt <- speed_filt(data, annotate = TRUE, verbose = TRUE, detailed_verbose = TRUE)

# setting that part of clean_track to FALSE, to annotate with other elements e.g. nsats
data_2 <- clean_track(data_speed_filt, speed_filt = FALSE, annotate = TRUE, hasgaps = TRUE) # using gaps already present in the data

# clean_track also gives a message if some satellite count data are found to be NA
# and these data are retained in the filter of number of satellites.

# See also processes for
#duplicate_track(data)
#tag_spans(data)
#filt_err(data)

# can also summarise the data, using a further function summary
#summary(data)

# see also 'raking' the data to check for potential outliers using a very basic quantile (here annotating)
data_test <- rake_outlie(data = data, var = "traj.speed", u_max = NULL, q = c(0.01,0.99), method = "quantile",
    plot = FALSE, annotate = TRUE, summary = TRUE, verbose = FALSE)

data_test$outlier_i_traj.speed_q  # outliers outside individual range
data_test$outlier_ia_traj.speed_q # outliers outside individual range + outside the all-animal range

# other approaches
trip::sda()

}
\seealso{
\link{rake}, @seealso \link{gap_section}, \link{filt_err}, \link{traj_speed}, \link{traj_speed_sm}, \link{speed_filt}, \link{turn_filt}

\link{read_track_MB}, \link{tspeed_jit}, \link{read_track_UvA}, \link{gap_section}, \link{summary}
}
