% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2coast.R
\name{dist2coast}
\alias{dist2coast}
\title{Distance of telemetry fixes to nearest mainland coastline}
\usage{
dist2coast(
  data,
  use_crs_shape = TRUE,
  shape = NULL,
  crop = TRUE,
  plot = FALSE,
  buffer_extent = 50000,
  p4s = 3035,
  verbose = TRUE
)
}
\arguments{
\item{data}{Input data object, with required columns: TagID, DateTime, longitude, latitude.}

\item{use_crs_shape}{logical defaulting to TRUE whether to use the projection of the shape provided for bird data, otherwise
the function will project both bird data and the shape using the p4s argument.}

\item{shape}{the function expects a single sf shape (sf MULTIPOLYGON or POLGON) representing the coastline.}

\item{crop}{logical defaulting to TRUE as to whether the polygon should be cropped by a square
extent (in association with the buffer_extent. This can greatly speed up the process
by not having the function attempt to compute nearest distances over an unnecessary large extent,
but the user \strong{MUST BE CONFIDENT} that the points bounded by the cropped shape outline will be those closest, and not
for instance a separate part of the polygon outside the crop, or the border of the actual crop. A warning
directs attention to this issue - it is best to plot first and carefully assess.}

\item{plot}{logical defaulting to TRUE giving a local pop up plot window.}

\item{buffer_extent}{A value to further add around the bird data to circumvent potential issues
of edges of the cropped map missing potential true land area; a default value is specified as 50000
but this will depend on map units and it is suggested this is checked before hand if using this option.}

\item{p4s}{The \code{sf} projection required for overlap processing, defaulting to BNG \code{epsg:3035}.}

\item{verbose}{logical, defaulting to TRUE, should detailed output messaging be returned on run of the function.}
}
\value{
The same dataset object as input with an additional column of dist2coast in metres.
}
\description{
Function for calculating the nearest distance to coast both offshore and inland using a specified polygon.
}
\details{
We may need to calculate how far animals may have travelled from a nearest coastline. Function \code{dist2coast}
fills that need, although this is a single GIS task obviously achievable in many ways outside this function.
}
\examples{

############### MAP PREPARATION ###############
# Idea here is to take a European map from mapdata and maps packages
# and combine with an even higher res map of UK and Ireland from personal libraries
# Using BNG projection for now, and having converted to sp format

# R packages not bundled with MoveRakeR
#remotes::install_github("ropensci/rnaturalearthhires")
#install.packages('rnaturalearth')
library(rnaturalearthhires)
library(rnaturalearth)

spdf_world <- ne_download(scale = 10, type = "countries")
Europe <- spdf_world[spdf_world$SUBREGION \%in\% c("Western Europe","Eastern Europe","Northern Europe","Southern Europe"),]
Europe_sf = st_crop(Europe, xmin = -25, ymin=32, xmax=35, ymax=70)
Europe_sf <- sf::st_transform(Europe_sf, 3035)
plot(Europe_sf$geometry)

data = data \%>\% clean_GPS() \%>\% dist2coast(use_crs_shape = TRUE, shape = Europe_sf,
    crop = TRUE, plot = TRUE, buffer_extent = 50000, p4s = 3035)

}
\seealso{
\link{offshore}, \link{trip_stats}
}
