% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounds.R
\name{get_bounds}
\alias{get_bounds}
\title{Calcualte the x and y min-max box ranges}
\usage{
get_bounds(data, p4s = 3035)
}
\arguments{
\item{data}{A \code{Track} family object, or just a simple data.frame or tibble.}

\item{p4s}{A numeric or \code{sf} \code{CRS} argument, here defaulting to the epsg string 3035.}
}
\value{
A list of length 2 with element one being the minimum and maximum x
range, and the second the y range.
}
\description{
This is a simple way of just getting the xy bounds of the data in \code{Track} datasets.
}
\examples{

indata <- yourdata # data.frame with a minimum of columns named TagID, DateTime, longitude, latitude
data <- Track(indata) # optional to convert to in-house class (most functions will still run on non track data but generic S3 methods will not be available for plot and summary)

# run get_bounds function
llyrb = get_bounds(indata, p4s = p4)

}
\seealso{
\link{plot.Track}
}
