% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TagIDs.R
\name{get_TagIDs}
\alias{get_TagIDs}
\alias{get_TagIDs.Track}
\title{Get unique TagIDs from Track, TrackStack or TrackMultiStack object}
\usage{
get_TagIDs(data, ...)

\method{get_TagIDs}{Track}(data, justlist = FALSE, TMS_join = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A \code{Track}, \code{TrackStack}, or \code{TrackMultiStack} object.}

\item{...}{further arguments passed to or from other methods.}

\item{justlist}{logical defaults to FALSE, to list all animals in the object
regardless if animals are repeated, e.g. across \code{TrackMultiStacks} and returns a charcter
vector for all names}

\item{TMS_join}{logical defaulting to TRUE specific to TMS objects whether to show TagID names in each component stack}

\item{verbose}{logical (default TRUE) to return detailed messaging.}
}
\value{
A vector containing unique TagIDs in the dataset(s).
}
\description{
Simple \code{unique()} application across listed formats supplied, useful for summarising
and assessing how to plot individual animals, e.g. using the S3 generic \code{plot} function for \code{Track} objects. The S3 method
for this class has only one function associated with it, being only useful for the Track-family.
}
