% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_accn.R
\name{is_Accn}
\alias{is_Accn}
\alias{is_AccnStack}
\alias{is_AccnMultiStack}
\alias{is_Accn2}
\title{is.Accn}
\usage{
is_Accn(x, strict = TRUE)

is_AccnStack(x, strict = TRUE)

is_AccnMultiStack(x, strict = TRUE)

is_Accn2(x, strict = TRUE)
}
\arguments{
\item{x}{Object to be assesed.}

\item{strict}{For is_ functions, logical for should the class check search strictly for the
specific sub-class, e.g. AccnStack, or AccnMultiStack (default = TRUE), but if FALSE
then the function checks as a normal "is" or inherits for the super class, e.g. Accn (even if object is a AS or AMS).}
}
\value{
Logical \code{TRUE} if \code{x} belongs to class \code{\link{Accn}}, \code{\link{AccnStack}} or \code{\link{AccnMultiStack}}, otherwise returns \code{FALSE}.
}
\description{
Check that an object is of class \code{\link{Accn}}, class \code{\link{AccnStack}} or class \code{\link{AccnMultiStack}}.
}
