% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_Track}
\alias{is_Track}
\alias{is_TrackStack}
\alias{is_TrackMultiStack}
\alias{is_Trip}
\alias{is_TripStack}
\alias{is_TripMultiStack}
\title{is_Track}
\usage{
is_Track(x, strict = TRUE)

is_TrackStack(x, strict = TRUE)

is_TrackMultiStack(x, strict = TRUE)

is_Trip(x, strict = TRUE)

is_TripStack(x, strict = TRUE)

is_TripMultiStack(x, strict = TRUE)
}
\arguments{
\item{x}{Object to be assesed.}

\item{strict}{For is_ functions, logical for should the class check search strictly for the
specific sub-class, e.g. TrackStack, or TrackMultiStack (default = TRUE), but if FALSE
then the function checks as a normal "is" or inherits for the super class, e.g. Track (even if object is a TS or TMS).}
}
\value{
Logical \code{TRUE} if \code{x} belongs to class \code{\link{Track}}, \code{\link{TrackStack}} or \code{\link{TrackMultiStack}}, otherwise returns \code{FALSE}.
}
\description{
Check that an object is of class \code{\link{Track}}, class \code{\link{TrackStack}} or class \code{\link{TrackMultiStack}}.
}
