% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_voltage.R
\name{is_Vo}
\alias{is_Vo}
\alias{is_VoStack}
\alias{is_VoMultiStack}
\alias{is_Vo2}
\title{is_Vo}
\usage{
is_Vo(x, strict = TRUE)

is_VoStack(x, strict = TRUE)

is_VoMultiStack(x, strict = TRUE)

is_Vo2(x, strict = TRUE)
}
\arguments{
\item{x}{Object to be assesed.}

\item{strict}{For is_ functions, logical for should the class check search strictly for the
specific sub-class, e.g. VoStack, or VoMultiStack (default = TRUE), but if FALSE
then the function checks as a normal "is" or inherits for the super class, e.g. Vo (even if object is a VS or VMS).}
}
\value{
Logical \code{TRUE} if \code{x} belongs to class \code{\link{Vo}}, \code{\link{VoStack}} or \code{\link{VoMultiStack}}, otherwise returns \code{FALSE}.
}
\description{
Check that an object is of class \code{\link{Vo}}, class \code{\link{VoStack}} or class \code{\link{VoMultiStack}}.
}
