% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Vo.R
\name{plot.Vo}
\alias{plot.Vo}
\title{Produce line plot of voltage data from a "Vo" object}
\usage{
\method{plot}{Vo}(
  data,
  dataTrack = NULL,
  start = NULL,
  end = NULL,
  roll_gap_thres = 3600,
  dt = FALSE,
  max_dt = NULL,
  dt_var = "vbat"
)
}
\arguments{
\item{data}{A \code{Vo} or \code{VoStack} object.}

\item{dataTrack}{An (optional, default NULL) A \code{Track} or \code{TrackStack} object for further plotting
of sunrise/sunset grey rectangles on plots and used also for getting sampling rates tags were
specified too using the DateTime stamp gaps between successive fixes (dt).}

\item{start}{Optional UTC start time (default NULL) for subsetted DateTimes (UTC) specified as a character
with format e.g.: "yyyy-mm-dd hh:mm:ss", although the function \code{MoveRakeR:datetime} is called
so "dd-mm-YYYY hh:mm:ss" and "dd/mm/YYYY hh:mm:ss" and two digit years should work too.}

\item{end}{As with start (default NULL) but the end UTC DateTime; note one or other of start and end
can be specified independently.}

\item{roll_gap_thres}{A numeric integer value for the maximum gap between nearest matched volateg data DateTimes
and the track data DateTimes, defaulting to 3600 s, and anything more than that will be dropped, i.e.
may be that the tag was not yet deployed, still recording voltage data but no GPS.}

\item{dt}{Logical (default FALSE) for whether the "dt" column in the Track data should be used to
plot in the secondary y-axis; only works if a Track-family object (TMS not accepted) is supplied
in the "dataTrack" argument.}

\item{max_dt}{The maximum dt value to retain int he plot (experimental) i.e. to remove all dt's greater than
this value; note lines are still drawn between gaps in the lineplot; only works if the
"dataTrack" argument is not NULL.}

\item{dt_var}{The variable to plot on the main y-axis in the plotting alongside "dt", defaulting
to the battery voltage (vbat) in the Vo dataset, but can accept the vsll solar charging data as well; only
works if the "dataTrack" argument is not NULL.}
}
\value{
A ggplot reflecting the user's inputs an wishes.
}
\description{
A generic plotting function \code{plot} available as an S3 method for objects of class \code{Vo}
and \code{VoTrack}.
}
\details{
The S3 generic \code{plot} function can be used for \code{Vo} and \code{VoStack} objects to
produce a visualisation of UvA-BiTS (not Movetech at the moment) tag voltages, to see
how tags are performing, e.g. just after deployment to check sustainability of sampling protocols.
The function in its default form
will accept a single \code{Vo} and \code{VoStack} object under the "data" argument, and
will then plot the "vbat" and "vsll" columns denoting battery and solar charging voltages
as blue and red lines respectively (secondary axis for solar voltage).

The function uses \code{ggplot} so each TagID supplied has its own facet.
But note, too many birds plotted at the same time will make this visualisation difficult.

The function also accepts "start" and "end" arguments (in UTC) if the user wishes
to subset the plot to be within specific DateTimes (one set across all birds only).

The option also exists to bring in a \code{Track} or \code{TrackStack} object
under the "dataTrack" argument, to work
alongside in the plot to produce rectangles of day and night (grey shaded) phases on the
plot using the R package \code{suncalc}.

Further, using the Track objects allows the user to also investigate the sammpling rate
of tags to see if particular sampling schedules were sustainable, with option then if
the argument dt = TRUE, to plot the sampling rate (dt) column in track data alongside either the
vbatt or vsll of the Vo data; the dt line will always appear on the secondary access as the
red line.

No doubt there could be bugs in this, so do let the package maintainer know if anything doesn't work as intended.
}
\examples{

db_file <- "GPS" # what you named the ODBC link on set up
db <- RODBC::odbcConnect(db_file) # currently required "db" to be named globally

TagIDs = c('1', '2', '3')
start = c("2021-06-21 06:00:00")
end = c("2021-07-25 09:00:00")
dataVo <- read_voltage_UvA(TagIDs, start=start, end=end) # form = energy is default
dataTrack <- read_track_UvA(TagIDs, start=start, end=end)

data = do.call('rbind',dataVo) # function will accept a TrackStack and VoStack as well....
dataTrack = do.call('rbind',dataTrack)

### plot simple battery vs solar charge, note hard to see with the timescale chosen
plot(dataVo)

### zoom in on the timeline
plot(dataVo, start = "2021-06-22 00:01:00", end = "2021-06-28 00:01:00")

### add in Track data for day/night rectangles#'
plot(data = dataVo, dataTrack = dataTrack, start = "2021-06-22 00:01:00", end = "2021-06-28 00:01:00")

### try a dt sampling rate plot
plot(dataVo, dataTrack, dt = TRUE) # note dt only works if a Track/TrackStack is supplied
plot(data = dataVo, dataTrack = dataTrack, dt = TRUE, start = "2021-06-22 00:01:00", end = "2021-06-28 00:01:00")

### see other options for toggling the thresholds of: max dt plotted, the max gap between the
### voltage data and nearest Track data on merge (roll_gap_thres),
### and plotting the solar charge instead of vbat default alongside dt using the dt_var variable.

}
\seealso{
\link{read_voltage_UvA}, \link{read_track_UvA}
}
