% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_pressure.R
\name{Pr}
\alias{Pr}
\alias{Pr2}
\alias{PrStack}
\alias{PrMultiStack}
\title{Pressure sensor ("Pr") constructor classes for MoveRakeR}
\usage{
Pr(data)

Pr2(data)

PrStack(data)

PrMultiStack(data)
}
\arguments{
\item{data}{An input base::data.frame for a pressure sensor dataset that is to be
coerced to the designated class.}
}
\value{
For tracking data, returns a \code{Pr}, \code{PrStack} or
\code{PrMultTrackStack}
}
\description{
Functions to append object classes of \code{Pr}, \code{PrStack} and \code{PrMultiStack},
obtained respectively using their respective namesake functions respectively. this is a specific
class for the UvA pressure sensor data.
}
\details{
Three special object classes are returned from initial the read in function
\code{\link{read_pressure_UvA}}. \code{Pr} is typically representative
of an individual animal's time-series, but an 'rbinded' \code{Pr}
object of several animals, i.e. multiple TagIDs, will also be accepted as a \code{Track}
object and processed for each individual, thus \code{Pr} objects represent a single \code{data.frame} class;
\code{PrStack} and \code{PrMultiStack} objects are nested \code{list} formats, and (as currently implemented)
\emph{must be treated as one individual per data.frame}. These stacked listed methods
are intended for efficient application of routines for groups of animals. Objects can be coerced into
\code{Pr}, \code{PrStack} and \code{PrMultiStack} classes but they require the following
minimum column names: \strong{TagID}, \strong{DateTime} and \strong{pressure_avg}.
Further checking of whether objects belong to special classes are provided by:
\code{link{is_Pr}},
\code{link{is_PrStack}} and \code{\link{is_PrMultiStack}}, and coercing of two
\code{PrStack} objects is provided by \code{\link{pressure_multi_stack}}.
}
\seealso{
\link{Track}, \link{Accn}, \link{pressure_multi_stack}, \link{is_Pr}, \link{is_PrStack}, \link{is_PrMultiStack}
}
