% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_pressure.R
\name{pressure_multi_stack}
\alias{pressure_multi_stack}
\title{pressure_multi_stack}
\usage{
pressure_multi_stack(data = NULL)
}
\arguments{
\item{data}{Object to be assesed.}
}
\value{
Return new \code{data} coerced to a \code{\link{PrMultiStack}} class.
}
\description{
Coerce \code{PrStack} class objects into a \code{PrMultiStack} class object.
}
