% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Track.R
\name{print.Track}
\alias{print.Track}
\title{Print S3 method for \code{Track} family objects}
\usage{
\method{print}{Track}(x, len = 6, ...)
}
\arguments{
\item{x}{An object of class \code{Track}, \code{TrackStack}, or \code{TrackMultiStack}.}

\item{len}{The desired length of the top and tail of the data to show, defaults to len=6 (integer value)
rows of \code{utils::head()} and \code{utils::tail()} respectively.}
}
\value{
No data are converted, only an on-screen "pretty" view is given.
}
\description{
A generic printing method for objects of class \code{Track}, \code{TrackStack} and \code{TrackMultiStack}.
}
\details{
Prints a simplified view of the data and a quick summary of the data structures in the R console terminal window,
following a combination of \code{data.table} and \code{tidyverse} implementations.
\code{print.Track} looks for columns "TagID","DateTime", "longitude", "latitude", for core printing
and then will display other columns of satellites_used and altitude, for now, although this
may change and expect flexibility to be increased. For \code{TrackStack} and \code{TrackMultiStack} objects,
the displayed data is given (top and tail) for the first Track object in the dataset.

Print options are also available under "options" as specified by functions: \code{\link{reset_options_mR}},
\code{\link{get_options_mR}} and \code{\link{set_options_mR}} to control the view of the printed output.
See help files for those functions.

Dev note: not yet coded flexibility should the first dataset not actually have enough data to print. Plus,
not yet coded warnings for differences in TS or TMS objects if there are differing numbers of columns
in list sub-datasets.
}
\examples{
indata <- yourdata # data.frame with a minimum of columns named TagID, DateTime, longitude, latitude
data <- Track(indata) # optional to convert to in-house class (most functions will still run on non track data but generic S3 methods will not be available for plot and summary)

print(data)

}
\seealso{
\link{get_options_mR}, \link{set_options_mR}, \link{reset_options_mR}
}
