% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_estimated2.R
\name{progress_estimated2}
\alias{progress_estimated2}
\title{Flexible progress bars with estimate time}
\usage{
progress_estimated2(total_steps)
}
\arguments{
\item{total_steps}{Integer. Total number of steps for the progress bar.}
}
\value{
A list with a \code{tick} function. Call \code{pb$tick(n = 1, label = NULL)}
to advance the progress bar by \code{n} steps and optionally display a label.
}
\description{
\code{progress_estimated2} is a replacement for dplyr::progress_estimated that was deprecated
in v 1.0.0 and inspired by that approach.
This is a simple light-weight progress bar that prints a single line progress
indicator with estimated remaining time. \code{progress_estimated2} also works well with nested loops and
optionally, a label can be used to indicate the current time ETA with print of the levels
from within loops.
}
\examples{
# Define a progress bar for a nested loop
tag_ids <- 1:3
vars_tag <- c("speed", "turn")
n_boot <- 5
gapsecs <- c(1,2)
seg_starts <- 1:4

total_steps <- length(tag_ids) * length(vars_tag) * n_boot *
               length(gapsecs) * length(seg_starts)
pb <- progress_estimated(total_steps)

for(ti in seq_along(tag_ids)){
  for(vi in seq_along(vars_tag)){
    for(b in seq_len(n_boot)){
      for(gap in gapsecs){
        for(i in seq_along(seg_starts)){
          Sys.sleep(0.05)  # simulate work
          pb$tick(label = sprintf("Animal \%d / Var \%s / Boot \%d", tag_ids[ti], vars_tag[vi], b))
        }
      }
    }
  }
}

# Or tick without a label
for(ti in seq_along(tag_ids)){
  for(vi in seq_along(vars_tag)){
    for(b in seq_len(n_boot)){
      for(gap in gapsecs){
        for(i in seq_along(seg_starts)){
          Sys.sleep(0.05)  # simulate work
          pb$tick()
        }
      }
    }
  }
}

}
