% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.R
\name{read_track_MB}
\alias{read_track_MB}
\title{Reading in animal tracking data from Movebank}
\usage{
read_track_MB(
  TagID = NULL,
  start = NULL,
  end = NULL,
  repo = NULL,
  option = "BTOTT",
  dropsat = FALSE,
  dropsats = 3,
  flt_switch = TRUE,
  mindata = 5,
  p4s = 3035,
  verbose = TRUE
)
}
\arguments{
\item{TagID}{The individual tag identifier from the Movebank database.}

\item{start}{The start time to delineate subsets of the total data for the individual in UTC.}

\item{end}{The end time to delineate subsets of the total data for the individual in UTC.
The start and end times must be given in the format "2014-06-25 15:20:23".
Arguments \emph{start} and \emph{end} default to NULL, reading in the entire track
of the animal and must be given in' \emph{character()} format, which are then converted to
\code{base::POSIXct} format within the \code{read_track_MB} function. Functionality is also
available for situations where only one or other of start and end need to be specified,
with the the other defaulting to NULL.}

\item{repo}{The Movebank repository where tracking data are stored. This is to
direct the function to look up animals and their data from Movebank in the right place.}

\item{option}{One of \code{"BTOTT"} or \code{"BTOTT_MB"} directing the function to return either a subsetted
aligned dataset in option "BTOTT" for combining with UvA data, or, "BTOTT_MB", containing the
required columns for displaying as a Track object in \code{MoveRakeR}, but doing no other changes
to the columns sourced from Movebank, i.e. cannot then be combined with UvA straight off; the latter is useful
if variables differ from those standard ones in MB or bespoke ones were added, or you have no idea
what UvA data is in which case this bit is probably meaningless, so just select option "BTOTT_MB". That said,
the "BTOTT" option will still search for missing column names that are needed for aligning with UvA
and if not found will insert NAs anyway BUT this may then get rid of further columns in MoveBank you want to keep.
The end result for both options is a \code{Track} object. If you don't want a \code{Track}, then just source
data from Movebank using \code{move::getMovebankLocationData}.}

\item{dropsat}{Logical defaults to FALSE. This indicates whether to drop GPS fixes with satellites below
the \emph{dropsats} threshold. It is suggested the user do not change these
from FALSE but instead use the function \code{\link{clean_GPS}}.}

\item{dropsats}{Numeric value indicating a threshold for satellite retention.}

\item{flt_switch}{logical argument, set to FALSE whether to retain
Movebank data erroneous flt_switch values, instead being dealt with in the
\code{\link{clean_GPS}} function.}

\item{mindata}{This specifies a minimum number of rows of data for the individual
to be retained and is a form of cleaning, but here is retained to exclude reading
in data from animals with too few data deemed unusable for analysis.
Defaults to mindata = 5. It can of course be set to zero if a completely raw dataset.}

\item{p4s}{A coordinate reference system argument using \code{sf::st_crs()}, thus the user just needs to enter in an epsg value, default = 27700.
The lat-longs extracted from the Movebank database will be projected via this CRS and stored in two new columns
\emph{X} and \emph{Y} (note capitilisation). Note however, that further functions
give options to change this if you wish in due course, and the user of course do all point-level manipulation
outside the functions if they so wish anyway. Argument Defaults to BNG using epsg string \code{CRS("+init=epsg:27700")}.}

\item{verbose}{logical (default TRUE) for whether detailed messaging is required.}
}
\value{
The final data outputted for individual animals is a \code{Track and data.frame} object.
The output of the function depends on the "option" selection. If "BTOTT" is chosen, then this will be
of 29 variables for each animal read in
with column names: Type, TagID, DateTime, speed_2d, speed_3d, traj_speed, latitude, longitude,
X, Y, dist, dt, satellites_used, gps_fixtime, gps.pdop, gps.hdop, altitude,
altitude_agl, h_accuracy, v_accuracy, pressure, temperature, x_speed, y_speed,
z_speed, speed_accuracy, direction, location and flt_switch.
It is assumed here the user will be familiar with most of these. But if option = "BTOTT_MB"
then the returned data will be whatever is stored in the Movebank database, but note, there may be
issues if combining TWO repos with DIFFERENT dataformats so beware in that instanace as it may
need individual repo read ins and further alignment.
For option "BTOTT", you will notice that for Movebank and data downloaded from the University of Amsterdam
i.e. from different sources, will have columns of all NAs for some variables
that were in one database but not another. This alignment has meant the
inclusion of missing variables in those cases. For example UvA has a gps.pdop
value but Movetech (for data used for initial function build) had gps.hdop.
Note also variables may differ to their
original truly "raw" naming convention in stored databases. The output for option "BTOTT" here is
to align and simplify for further analysis. All time units are in seconds
and distances in metres, speeds in m/s. Importantly "location" refers to the
individual row.name identifier in respective databases, so that no matter what
data are extracted in this function, the saved data can be matched back and if
necessary recreated through the location variable if need be. This is important
for data that is continually being updated remotely and particularly if you
are working with the full data downloaded from animals up to very present.
Data at the individual level are always provided as a \code{Track} class, and for multiple
TagIDs, a \code{TrackStack} object is returned. Note, read in functions do not currently support
\code{TrackMultiStack} objects (i.e. the higher level listed class of \code{TrackStack} objects).
For \code{TrackMultiStack} objects, \code{TrackStack} objects should be concatenated and coerced using \code{\link{track_multi_stack}}.
Names of each animal are provided for \code{Track} lists.
**** NOTE TO PACKAGE MAINTAINER: Need to add flexibility in this naming, e.g. should
multiple breakdowns of the same animal be needed, or if TagIDs are duplicated in different databases.
}
\description{
Connect and download tracking data directly from Movebank repositories.
}
\details{
At the core of this function is the desire to have wrapped flexibility
to extract sections of data across multiple individuals from specific
repositories. \code{read_track_MB} is therefore designed so that a "control file"
i.e. with static start and end date-times for TagIDs can be used as a lookup table
together with the repository to extract the necessary data with ease.

\code{read_track_MB} is built around R package \code{move}. The user may need to
access repositories that are not public and so will need the necessary
privileges for repositories they are trying to access. The use of
\code{read_track_MB} also assumes that the user has specified their login in the
global environment. The function makes calls to  \code{move::getMovebankAnimals}
and \code{move::getMovebankLocationData} to extract data. WARNING: \code{read_track_MB}
makes it possible to extract large amounts of data. However, it is therefore advisable
to make as few direct calls to the databases and extract the data needed to be stored
and then read in again for subsequent analytical steps.

NOTE: \code{read_track_MB} has recently been
updated to improve speed and efficiency; it was previously making repeated calls to
the \code{move} package, but now the function downloads all data requested and subsets time periods
afterwards within the routine. This also allows a little more flexibility for \code{read_track_MB} such that the length of
"repos" supplied does not need to be neatly balanced to each animal supplied, e.g. can be multiple TagIDs
and one repo. Also NULL can be used for start and ends (as is default) as indicated below,
which just downloads all data for that TagID from the repo. Should TagIDs need different start/ends
these must be balanced, so length(TagID) must equal length(start) and length(end), but character-format
"NULL" can be used within that bird-specific for start and ends i.e. similar to above.

\code{read_track_MB} along with it's sister function \code{\link{read_track_UvA}}
are designed to extract GPS positional data on the fly in an efficient manner and to directly coerce
the resultant \code{base::data.frames} objects to \code{Track}, \code{TrackStack} and \code{TrackMultiStack} classes.
The user can also specify \emph{start} and and \emph{end} DateTimes (see below)
to specify subsections from the entire dataset available in repositories for individual
animals. R library \code{adehabitatLT:as.:traj} is also called to add a trajectory speed to the dataset
between fixes, along with a "dt" time difference (secs) and distance (m) between
fixes. Note that occasionally the UvA database can go down, and recently the RODBC connection
was not working on multiple computers. It is best to email the Dev, or email \code{UvA-BiTS"at"uva.nl}.

There are other arguments included in this function for cleaning, but these have
been transferred to the function \code{\link{clean_GPS}} function although they
are retained here are set to FALSE (see Arguments).
}
\examples{
login <- move::movebankLogin("username","password")
repo = "repo" # Your MB repository name
TagID = c("A","B","C")
start = c("2016-06-01 13:53:50", "2016-06-15 12:22:13", "2016-06-05 08:07:23")
end = c("2016-07-15 09:17:14", "2016-07-20 01:08:58", "2016-07-18 14:22:45")

dataMB <- read_track_MB(TagID,repo=repo,start=start,end=end)

# example also reading in full datasets with NULL start and end times defined.
dataMB2 <- read_track_MB(TagID,repo=repo) # testing reading in all the data of animals with no specific start ends (NULL be default)

}
\references{
Kranstauber, B. Smolla, M. & Scharf, A.K. (2020). move:
Visualizing and Analyzing Animal Track Data. R package version 3.3.0.
\url{https://CRAN.R-project.org/package=move}
}
\seealso{
\link{read_track_UvA}, \link{Track2move}
}
