% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{reset_options_mR}
\alias{reset_options_mR}
\alias{set_options_mR}
\alias{get_options_mR}
\title{Global options for printing \code{Track} datasets}
\usage{
reset_options_mR()

set_options_mR(Print_method = "Track", Print_type = "subset")

get_options_mR()
}
\arguments{
\item{Print_method}{One of either \emph{"Track"} or \emph{"data.frame"} to determine printing as for \code{Track} family
layout or the \code{base data.frame} view. For \code{Track} classes, this is the default option,
and indicates the print method dispath will display as a coloured layout summarising the data
contained in the dataset, i.e. rows, columns, and TagIDs. For the \code{data.frame} layout, this is
as it is in the normal printing of data.frames. For \code{TrackStack} and \code{TrackMultiStack}
listed formats, selecting the \code{data.frame} \emph{Print_method} will display as you would normally
see for listed formats in base R.}

\item{Print_type}{One of either \emph{"subset"} or \emph{"all"} controlling the extent of data
to be displayed. If "subset" is chosen, for both "Track" and "data.frame" \emph{print_type} options,
the data will be restricted to the top rows of the dataset as controlled in the \emph{len} option in the
\code{print.Track} function. For "subset" chosen with the "Track"  \emph{Print_method}, this will result
in core column printing from the Track-dataset provided, with abbreviated display of
TagID, lat-longs and further satellite and altitude columns.
If "all" is chosen, then all columns are displayed unabbreviated. No abbreviations are made if
"subset" is chosen with the "data.frame" \emph{Print_method}; instead, this will simply result in
a restricted display of rows via the \emph{len} option in the \code{print.Track} function.}
}
\value{
Alterations to the \code{R6} class \code{options_mR} for printing.
}
\description{
Options to control how \code{Track} datasets are displayed in the console.
}
\details{
By default, \code{Track} family datasets (i.e. Track, TrackStack and TrackMultiStack classes)
will print with a specific colour format for each class type, displaying limited columns
and giving information of other columns not displayed; this follows a tibble-style of formatting.
However, the user may wish to disable this and revert back to e.g. a \code{data.frame} style,
or the full columns of data may wish to be displayed.

Global options have therefore been coded, using the \code{R6} class that is loaded as \code{options_mR}.
The user should not edit this directly. Instead, there are three functions: \code{set_options_mR},
\code{get_options_mR} and and \code{reset_options_mR}.

\code{set_options_mR} includes two arguments for controlling the \emph{Print_method} and \emph{Print_type},
and \code{reset_options_mR} resets these to the defaults of "Track" and "subset", respectively,
indicating a "Track" coloured summary layout and a "subset" view of just core columns (see details).
In other words, \emph{Print_method} controls the overall viewing style of the print, whereas \emph{Print_type} controls
the column and row length layout. The function \code{get_options_mR} simply displays the current
options.
}
\examples{
library(MoveRakeR)

# check the current (default) options:
get_options_mR()

# switch to a full print of the Track-family class
set_options_mR(Print_method = "Track", Print_type = "all")
get_options_mR()

# if the data are a stack of animals these can be viewed as:#'
print(data[[1]][[1]]) # first "Track" class of the TMS
print(data[[1]]) # first stack summary
print(data) # full TMS summary


# reset back to defaults
reset_options_mR()
get_options_mR()

# switch to base R standard data.frame printing
set_options_mR(Print_method = "data.frame", Print_type = "subset")
get_options_mR() # with subset

print(data[[1]][[1]], len = 10) # i.e. prints "len" number of top rows (len defaults to 6)
print(data[[1]], len = 10) # listed formats retained as would be displayed normally
print(data, len = 3) # as above

# reset back to defaults
reset_options_mR()

# if wanting to revert to full original way of base R printing track objects:
set_options_mR(Print_method = "data.frame", Print_type = "all")

print(data[[1]][[1]]) # "len" argument now makes no difference as bypassed

# wouldn't advise printing this (tries to print everything up to max getOptions base R "max.print" options):

reset_options_mR()

### END

}
\seealso{
\link{print.Track}
}
