% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Method to simplify Track objects to essential data}
\usage{
simplify(data, keep = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Input data object, of class \code{data.frame}, \code{tibble-family} or \code{data.table}, with required columns: TagID, DateTime, longitude, latitude.}
}
\description{
Simplification of telemetry dataset to retain TagID, DateTime, longitude, and latitude columns,
plus any extra column the user wishes.
}
