% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trip.R
\name{simulate_trip}
\alias{simulate_trip}
\alias{sim_data}
\title{Simulate animal movement trips with correlated random walks}
\usage{
simulate_trip(
  n = 50,
  hub_lat = 40,
  hub_lon = 0,
  rho_explore = 0.94,
  rho_forage = 0.4,
  rho_return = 0.98,
  step_frac_explore = 0.7,
  step_frac_forage = 0.3,
  step_frac_return = 1,
  step_sd_explore = 0.2,
  step_sd_forage = 0.5,
  step_sd_return = 0.3,
  angle_noise_explore = pi/15,
  angle_noise_forage = pi/1,
  angle_noise_return = pi/20,
  min_return_steps = 10,
  interpolate_equal_space = FALSE,
  dt = 300,
  max_speed = 50
)

sim_data(
  animals = c("A001", "B002", "C003"),
  n_trips = 5,
  points_per_trip = 50,
  hub_lat = 40,
  hub_lon = 0,
  dt = 300,
  tol = 0.05,
  cp_time_mean = 6,
  cp_time_sd = 1,
  points_variation = 0.3,
  cp_time_jitter = 0.5,
  max_speed = 50,
  seed = NULL,
  min_return_steps = 10,
  default_ranges = NULL,
  ...
)
}
\arguments{
\item{n}{Integer. Total number of steps in the simulated trip (excluding any additional steps added for return to the hub), within \code{sim_data()}, default = 50.}

\item{hub_lat}{Numeric, latitude of the central hub, within \code{sim_data()} and \code{simulate_trip()}, default = 40.}

\item{hub_lon}{Numeric, longitude of the central hub, within \code{sim_data()} and \code{simulate_trip()}, default = 0.}

\item{rho_explore}{Numeric between 0 and 1. Directional autocorrelation for the exploratory (outbound) phase. Higher values produce straighter paths, default 0.94 within \code{simulate_trip()}.}

\item{rho_forage}{Numeric between 0 and 1. Directional autocorrelation for the area-restricted search / foraging phase. Lower values produce more tortuous paths, default 0.4 \code{simulate_trip()}.}

\item{rho_return}{Numeric between 0 and 1. Directional autocorrelation for the return / commuting phase. Higher values produce straighter paths toward the hub, default 0.98, \code{simulate_trip()}.}

\item{step_frac_explore}{Numeric. Fraction of maximum step length to use in the exploratory phase. Controls relative movement speed/length, within \code{simulate_trip()}, default = 0.7.}

\item{step_frac_forage}{Numeric. Fraction of maximum step length to use in the foraging phase. Usually smaller to simulate local search behavior, within \code{simulate_trip()}, default = 0.3.}

\item{step_frac_return}{Numeric. Fraction of maximum step length to use in the return phase. Usually larger to generate straighter, faster commuting movement, within \code{simulate_trip()}, default = 1.0.}

\item{step_sd_explore}{Numeric. Fractional SD for step length in exploratory phase (step length multiplier), within \code{simulate_trip()}, default = 0.2.}

\item{step_sd_forage}{Numeric. Fractional SD for step length in foraging phase (step length multiplier), within \code{simulate_trip()}, default = 0.5.}

\item{step_sd_return}{Numeric. Fractional SD for step length in return phase (step length multiplier), within \code{simulate_trip()}, default = 0.3.}

\item{angle_noise_explore}{Numeric. Standard deviation of angular noise (radians) in the exploratory phase; controls path curviness, within \code{simulate_trip()}, default = pi/15.}

\item{angle_noise_forage}{Numeric. Standard deviation of angular noise (radians) in the foraging phase; controls path curviness, within \code{simulate_trip()}, default = pi.}

\item{angle_noise_return}{Numeric. Standard deviation of angular noise (radians) in the return phase; controls path curviness, within \code{simulate_trip()}, default = pi/20.}

\item{min_return_steps}{Integer. Minimum number of steps to complete the return phase, regardless of distance to the hub, within \code{simulate_trip()}, default = 10.}

\item{interpolate_equal_space}{Logical. If TRUE, additional points are interpolated to produce a track with more uniform spacing, within \code{simulate_trip()}, default = FALSE.}

\item{dt}{Numeric. Time step between successive locations, in seconds. Used for approximate speed and DateTime timestamp calculation, within \code{sim_data()} and \code{simulate_trip()}, default = 300 s.}

\item{max_speed}{Numeric, maximum speed in meters/second; used to constrain movement, within \code{sim_data()} and \code{simulate_trip()}, default = 50 m/s.}

\item{animals}{Character, concatenated string for ID of animals, for \code{sim_data()}, default of three animals "A001", "B001" and "C001".}

\item{n_trips, }{Integer, number of trips to simulate, default = 5, within \code{sim_data()}.}

\item{points_per_trip}{Integer, for number of points per trip to simulate, default 50, within \code{sim_data()}.}

\item{tol}{Numeric, relative jitter for dt per fix to simulate irregular tag sampling (0 = fixed dt), used within \code{sim_data()}, default = 0.05.}

\item{cp_time_mean}{Numeric, mean central-place stay duration in hours between trips, default = 6 hours, within \code{sim_data()}.}

\item{cp_time_sd}{Numeric, standard deviation of central-place stay duration in hours, default = 1, within \code{sim_data()}.}

\item{points_variation}{Numeric (0–1), proportional variation in points per trip, default = 0.3, within \code{sim_data()}.}

\item{cp_time_jitter}{Numeric (hours), additional random jitter added to CP stay times, default = 0.5, within \code{sim_data()}.}

\item{seed}{Integer for seed to set for reproducibility, within \code{sim_data()}, default = NULL (random seed).}

\item{default_ranges}{Named list of numeric vectors of length 2 specifying min and max values for trip-level parameters.
Each trip randomly samples a value within the range for each parameter. Defaults are:
\describe{
\item{rho_explore}{c(0.6, 0.8) directional persistence during exploration phase.}
\item{rho_forage}{c(0.2, 0.5) directional persistence during foraging phase.}
\item{rho_return}{c(0.8, 0.99) directional persistence during return phase.}
\item{step_frac_explore}{c(0.6, 0.8) proportion of max step length used in exploration.}
\item{step_frac_forage}{c(0.2, 0.4) proportion of max step length used in foraging.}
\item{step_frac_return}{c(0.9, 1.0) proportion of max step length used in return.}
\item{angle_noise_explore}{c(pi/20, pi/4) angular noise (radians) during exploration.}
\item{angle_noise_forage}{c(pi/2, pi) angular noise (radians) during foraging.}
\item{angle_noise_return}{c(pi/50, pi/10) angular noise (radians) during return.}
}
Any parameters omitted from the list will fall back to these defaults. Any parameters omitted from the list
will fall back to default ranges and dach trip randomly samples a value within the range for each parameter.}

\item{...}{Additional parameters passed to \code{simulate_trip()} from \code{sim_data()}, e.g.,
\code{rho_explore}, \code{rho_forage}, \code{rho_return}, \code{step_frac_explore},
\code{step_frac_forage}, \code{step_frac_return}, \code{angle_noise_explore},
\code{angle_noise_forage}, \code{angle_noise_return}, \code{min_return_steps}.}
}
\value{
From \code{simulate_trip()}, a \code{data.frame} with two columns:
\describe{
\item{latitude}{Simulated latitude positions.}
\item{longitude}{Simulated longitude positions.}
}

From \code{sim_data()}, A tibble (or \code{Track} object if \code{Track()} is applied) containing:
\describe{
\item{TagID}{Animal identifier.}
\item{trip}{Trip number.}
\item{DateTime}{POSIXct timestamp of each fix.}
\item{longitude}{Longitude of the point.}
\item{latitude}{Latitude of the point.}
\item{dist_m}{Distance (m) from previous point.}
\item{dt_s}{Time difference (s) from previous point.}
\item{speed_m_s}{Observed speed (m/s) between consecutive points.}
\item{pdop}{Simulated GPS PDOP (1–9).}
\item{nsats}{Simulated number of satellites (3–12).}
\item{speed_basic}{Binary speed flag (0 = <300 m/s, 1 = >=300 m/s).}
}
}
\description{
These functions simulate artificial movement trajectories of animals departing from and returning to a
central hub (e.g. a colony or roost), using correlated random walk models.
}
\details{
The function \code{simulate_trip()} generates a trajectory to simulate GPS data collected from animals,
with distinct behavioral phases (exploration, foraging, return) and adaptive return dynamics.
Step lengths and turning correlations differ by phase, and the return phase meanders
naturally towards the hub and extra points are added if needed to reach the hub.

\code{simulate_trip()} also incorporates a duration of time at the central place between trips, and randomises the
components related to number of points per trip (points_per_trip), DateTimes (dt), and the measures of straightness and
angle noise for each of the distributions for exploratory, foraging and return commute phases. These choices can be customised.
}
\examples{
# Simulate 100 trips for three animals with default parameters
fake_data <- simulate_trip(n = 100, hub_lat = 40, hub_lon = 0)

plot(fake_data$longitude, fake_data$latitude, type = "l", col = as.numeric(as.factor(fake_data$TagID)))
#fake_data[fake_data$TagID == "A001",]
#fake_data \%>\% group_by(TagID, trip) \%>\% summarise(start_trip = min(DateTime))
#fake_data[fake_data$TagID == "A001",]

plot_leaflet(fake_data)


}
