% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Track.R
\name{[.Track}
\alias{[.Track}
\title{Subsetting for Track objects}
\usage{
\method{[}{Track}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A Track object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{drop}{Logical, whether to drop dimensions}
}
\description{
Ensures Track class is preserved when subsetting.
}
