% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_calc.R
\name{sun_calc}
\alias{sun_calc}
\title{Add parameters from R package \code{suncalc} to data}
\usage{
sun_calc(data, keep = c("sunrise", "sunset"))
}
\arguments{
\item{data}{A \code{Track} family, \code{tbl}, or \code{Track2} object.}

\item{keep}{Corresponding parts of the returned \code{suncalc} outputs to retain; see
the \code{suncalc} help files for more information; specified as a concatenation of character vectors
defaults to c("sunrise","sunset"). Note, this "keep" is for sub-setting of the Track
object and by default all variables are computed in \code{suncalc} to estimate
finest level categorisation of points by twilight grouping. Options of "all"
can also retain all the variables from \code{suncalc} and "none" drops all, after
categorisation of fixes.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The same Track-style dataset with additional columns retained from the specification of the
keep argument. A further column of \strong{"dnt"} is included with levels of
\itemize{
\item DA, full daylight
\item NI, full night
\item AT, Astronomical twilight
\item NT, Nautical twilight
\item CT, Civil twilight
\item SR, Sunrise (between sun appearing on horizon and to fully above)
\item SS, Sunset, (between sun first touching horizon to fully below)
\item suffixes of "m" and "e" are added to "AT", "NT" and "CT" denoting the side of the day, "morning" or "evening"
\item \emph{Note}: Astronomical twilight often ends up as NA, meaning that points after nautical
twilight will be placed in the "NI" night category, rightly or wrongly! There will likely be
quirks with this. However, of course the user can return the columns and make bespoke
calculations if need be retaining columns from \code{suncalc} using the "keep" argument.
}
}
\description{
A wrapper calling R package \code{suncalc} to add needed parameters of sun position
sunlight phase, moon positions and lunar phases.
}
\details{
Although the process of using suncalc is super easy and can be done by the user as they wish, if you are after
swift easy calculation of simple (e.g. sunrise, sunset) values then this function
just does it for you, rather than applying afresh each time with user-specific code.
}
\examples{

# this should be installed already on load in of MoveRakeR:
#install.packages("suncalc")
#library(suncalc)

data <- sun_calc(data, keep = c("sunrise", "sunset"))
table(data$dnt)

}
\references{
Thieurmel, B. & Elmarhraoui, A. (2019). suncalc: Compute Sun Position, Sunlight Phases, Moon Position and Lunar Phase. R package version 0.5.0. \url{https://CRAN.R-project.org/package=suncalc}
}
