% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_trips.R
\name{tabulate_trips}
\alias{tabulate_trips}
\title{Tabulated summary of trip statistics for report presentation}
\usage{
tabulate_trips(
  data,
  roundval = 2,
  TD_unit = "days",
  by = NULL,
  rm_inc = TRUE,
  option = "means"
)
}
\arguments{
\item{data}{A \code{Trip} object, from function \code{\link{trip_stats}}.}

\item{roundval}{Numeric integer for the round value of decimal places for summary metrics, set to 2 as default.}

\item{TD_unit}{One of three options for the units of trip duration; defaults to 'days'
but the user can specify 'mins' or 'hours'.}

\item{by}{Additional variables to summarise data over, such as year perhaps, or stage of season,
which need to be present in the \code{Trip} data supplied. Means for 'all' animals are also estimated
for all animals, as well as for the 'by' grouping (defaults to NULL).}

\item{rm_inc}{Logical defaulting to TRUE, for whether incomplete trips should be removed
from statistics.}

\item{option}{a character vector of either \code{"means"} or \code{"boxplots"}; these options are for how
the user wants the metrics summarised, either as means+/-SD and (max) value, for the "means" option,
or, probably more appropriately, presenting as median values of the distribution (upper and lower quartile presented in parentheses),
and max value presented in square brackets, for the "boxplots" option. Obviously see \code{graphics::boxplot} on which these
parameters are extracted.}
}
\value{
A \code{data.frame} of several variables, depending on those available and elected by the user,
from earlier choices made in the \code{trip_stats} function.
The function will automatically sumarise core variables of
\itemize{
\item \code{Distmax}, representing the maximum distance of distal point
\item \code{DistTotal}, the total distance traveled on the trip, and
\item \code{TripDur}, and the time elapsed from start to finish, respectively.
All units of distance are in km and time in hours.
Also outputted in all instances of use of the function are:
\item \code{TagID}, for the animal identifier,
\item \code{n}, number of trips for number of trips the animal made,
\item \code{incompl} and the number of "incomplete" trips identified via gapsections present.
If offshore and inshore distinctions have been used in the' \code{trip_stats} function, then
additional columns are processed to give:
\item \code{DistOff} representing animal-level summaries of maximum distances traveled per trip (distal point)
within offshore areas (as delineated by function \code{offshore}).
\item \code{DistOn}, the same as above for inshore areas.
\item \code{n_off} Number of trips offshore.
\item \code{n_on} Number of trips onshore.
\item \code{incompl_off} Number of incomplete trips offshore.
\item \code{n_compl_off} Number of complete trips offshore.
\item \code{incompl_on} Number of incomplete trips onshore.
\item \code{n_compl_on} Number of complete trips onshore.
Further, should distance-to-coast information be present within input Trip classes fed to
\code{tabulate_trips},' then the function will summarise additional variables in the returned output:
\item \code{DistCoastOff}, the maximum distance from the coast for offshore strata, and
\item \code{DistCoastOn},  the same as above for inshore areas.
}
}
\description{
\code{tabulate_trips} is a simple summary function to tabulate information on 'trips' of
animals from a central place. This table follows on from initial annotating of telemetry
data as trips through function \code{define_trips} and subsequent running of function
\code{trip_stats} to produce trip-level information per animal.
}
\details{
The sole purpose of this function is to make reporting trip statistics easier,
and repeatable. The outputs from this function are contingent on other choices
made throughout the data processing. Also output in the function output is a line representing "All" birds,
including for additional selection of other variables in the "by" argument.
Note, these 'all' columns are arithmetic means across birds thus, birds with more trips will contribute more to this value.
Options for offshore/onshore delineation and distances reached inland and offshore from the
coast (stemming from choices in \code{trip_stats}) are also provided.
}
\examples{

ColLon = -3.185
ColLat = 54.045
p4 <- sf::st_crs((paste("+proj=laea +lon_0=", ColLon," +lat_0=", ColLat, " +units=m", sep="")))

indata <- yourdata # data.frame with a minimum of columns named TagID, DateTime, longitude, latitude
indata <- Track(indata) # optional to convert to in-house class (most functions will still run on non track data but generic S3 methods will not be available for plot and summary)

trip_test <- indata \%>\% clean_GPS() \%>\% # 8 hour gapsections by default
  define_trips(method="rect", lls = c(-3.2, 54.0553,-3.1689, 54.0437), p4s = p4) \%>\%
  trip_stats()

# tabulate trips by year
tabulate_trips2(trip_test, by = "year")

# or just overall
tabulate_trips2(trip_test)

# This function could easily be better generalised in future if more variables from trip stats were added, e.g. tortuosity etc

}
\seealso{
\link{trip_stats}, \link{define_trips}
}
