% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_spans.R
\name{tag_spans}
\alias{tag_spans}
\alias{draw_bbox}
\title{General summary of data quality extent in space and time}
\usage{
tag_spans(
  data,
  measure = c("general", "spatial", "temporal"),
  xra = c(-10, 5),
  yra = c(25, 65),
  dates = c("2014-01-01 00:00:01", "2021-12-31 23:59:59"),
  buffer_days = 2,
  tz = "UTC",
  max_show = 3,
  messages = TRUE,
  msg_col = 33
)

draw_bbox(
  width = 40,
  height = 15,
  col = "35",
  xmin = -10,
  xmax = 10,
  ymin = -5,
  ymax = 5,
  title = "Spatial extent (long, lat)"
)
}
\arguments{
\item{data}{Input data object, with required columns: TagID, DateTime, longitude, latitude.}

\item{measure}{One of three options, "general", "spatial", and/or "temporal", with any combination allowed. This
directs the function to assess the (1) general data quality using the \code{summary} function in \code{moveRakeR},
being the overall bounding xy box, within which animals had most and least fixes, which animals were
at the extremes of the bounding box, which animals were
tracked longest and for the least amount of time; (2) the spatial quality, i.e. any fixes outside
and which birds had such fixes occuring outside the supplied xra and yra, and (3) temporal quality being either an errant DateTime in the future beyond the current date or outside
bounds of supplied DateTimes.
a data in the past.}

\item{xra}{The expected limits of acceptable data in the x longitude dimension (WGS84),
supplied as a two-length vector c(min,max), c(-10,5).}

\item{yra}{The expected limits of acceptable data in the y latitude dimension (WGS84),
supplied as a two-length vector c(min,max), defaulting to c(25,65).}

\item{dates}{The expected valid range of dates over which the dataset was collected, supplied as a
two-length vector c(min,max) as character YYYY-MM-DD HH:MM:SS format, defaulting to c("2014-01-01 00:00:01", "2021-12-31 23:59:59").}

\item{buffer_days}{A buffer in days for flagging dates as a temporal flag either side of the 'dates' time window.}

\item{max_show}{The maximum number of individual animals to show in the messaging print out.}

\item{messages}{Whether printed console messages are to be returned to the user, a Boolean defaulting to TRUE.}

\item{msg_col}{A numeric value for the colour to be used in console graphics.
The colour of the print out in the console, depending on your taste and whether black or white backgrounds
are currently being used within R Studio. Defaults to 33 (yellow).}
}
\value{
An object as a list, length = 4, depending on the inputs to the function for the measure, xra, yra, and dates
arguments. A full spatio-temporal output would be obj[\link{1}], for the general data summary, using the
\code{summary} function, obj[\link{2}] for the xra summary, obj[\link{3}] for th yra summary and obj[\link{4}] for the date range summary.
Each summary is returned as a \code{tibble} with TagID, and then specific columns for each summary. The first object
is the same output as returned through the \code{summary} function - see that documentation.

For xra obj[\link{2}],
columns are:
\itemize{
\item any_xmin_bad: a Boolean of whether TagIDs had any fixes below the minimum longitude;
\item any_xmax_bad: a Boolean of whether TagIDs had any fixes above the maximum longitude;
\item xmin_bad: a list of which values were below the minimum longitude;
\item xmax_bad: a list of which vallues were above the maximum longitude;
\item xmin_bad_times: a list of DateTimes in order for which values were below the minimum longitude;
\item xmax_bad_times: a list of DateTimes in order for which values were above the maximum longitude.
}

The same output is returned for latitude obj[\link{3}].

For the fourth obj[\link{4}] for the temporal summary, columns are given for:
\itemize{
\item any_future: a Boolean for whether TagIDs had any fixes found to be in the future (past the DateTime clock from when the function wss run);
\item future_dates: a list of future DateTimes if any were outside the current DateTime;
\item n_future: the number of future DateTimes found;
\item any_temporal: a Boolean for whether TagIDs had any DateTimes before or after the supplied 'date' range;
\item temporal_dates: a list of DateTimes that were outside the date range;
\item n_temporal: the number of fixes outside the date range.
}
}
\description{
Summarises the spatial and temporal range of the data and perceived outliers, building on that of summary.Track
}
\details{
The \code{tag_spans} function highlights potentially errant spatial and/or temporal data falling outside of
user provided bounds of data, and is therefore a 'soft' rake of the data, i.e. not formally identifying outliers.
Messaging is provided back to the user to indicate the spatio-temporal range and which animals in the data
have potentially outlying GPS fixes. This is useful if the user knows the overall extent of the area used,
or has an approximation of it, and therefore sudden jumps in the GPS can be identified easily. Note also,
these jumps are the very same fixes that would be identified under the \code{speed_filt} function.

Note that the xra and yra arguments are only a guide of what you may expect to be incorrect.
This is an intentionally simplistic first pass of the data.
}
\examples{

dat <- tag_spans(data, measure = c("general","spatial", "temporal"),
    msg_col = 33,
    xra = c(-10,5), yra = c(25,65), max_show = 3,
    dates = c("2014-01-01 00:00:01", "2021-12-31 23:59:59"), buffer_days = 2,
    messages = TRUE)

length(dat) # 4
dat[[2]] # returned summary of assessment for xra criterion

}
\seealso{
\link{rake}, \link{rake_outlie}, \link{tag_timeline}
}
