% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_voltage.R
\name{Track2Vo}
\alias{Track2Vo}
\title{Track2Vo}
\usage{
Track2Vo(data, vsll = NULL, vbat = NULL)
}
\arguments{
\item{data}{Track object to be converted.}

\item{vsll}{column in your data relating to solar voltage charging (e.g. voltage or percentage)}

\item{vbat}{column in your data relating to the battery charge (e.g. voltage, or percentage)}
}
\description{
Conversion between \code{Track} and \code{Vo} classes
}
\details{
The \code{Track2Vo} function is required for use of the plot generic S3 method for \code{Vo} objects,
to first create a \code{Vo} object from you track data based on columns that are assumed to be existing
within the \code{Track} data.
}
\examples{
Vo_obj = Track2Vo(data, vbat = "tag-voltage", vsll = "battery-charging-current")

Vo_obj_subset = Vo_obj[Vo_obj$TagID == "animal_1",] # subsetting for one animal if wanted
track_data = data[data$TagID == "animal_1",]

# plot with track data alongside for additional day/night bars, with a subset of time period
plot_st = plot(Vo_obj_subset, dataTrack = track_data, start = "2022-04-05 00:00:00", end = "2022-04-10 00:00:00")

}
