% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{track_multi_stack}
\alias{track_multi_stack}
\alias{track_stack}
\alias{trip_multi_stack}
\title{track_multi_stack}
\usage{
track_multi_stack(data = NULL)

track_stack(data = NULL)

trip_multi_stack(data = NULL)
}
\arguments{
\item{data}{Object to be assesed.}
}
\value{
Return new \code{data} coerced to a \code{\link{TrackMultiStack}} class.
}
\description{
Coerce \code{TrackStack} class objects into a \code{TrackMultiStack} class object.
}
