% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_filt.R
\name{turn_filt}
\alias{turn_filt}
\title{Trajectory turning angle filter for GPS data}
\usage{
turn_filt(
  data,
  turnFilt = pi/4,
  turnFilt_dir = c("less_than", "greater_than"),
  annotate = FALSE,
  hasgaps = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Input data object, with required columns: TagID, DateTime, longitude, latitude.}

\item{turnFilt}{Numeric value in radians below or above which to filter or annotate data, defaults to  pi/4.}

\item{turnFilt_dir}{The direction above or below the turnFilt value for which to annotate or filter; one of "less_than" or "greater_than", defaults to former.}

\item{annotate}{TRUE or FALSE (default) for whether to annotate (TRUE) or filter (FALSE) the data by the turnFilt condition.}

\item{hasgaps}{TRUE (default) or FALSE whether to check for existing gaps having been defined already in the data,
and if found use those gaps, if not reverting to \code{gap_section} defaults; FALSE also takes the \code{gap_section} defaults.}

\item{verbose}{TRUE (default) or FALSE whether to provide within-function messaging.}
}
\value{
Either an annotated dataset with an additional column of 'angle_rm' (1,0) for whether data were 'flagged', or a filtered
dataset with such rows having been removed.
}
\description{
This function filters or annotates data for fixes that are greater than or less than a user-provided
turning angle
}
\details{
It may be necessary to flag/annotate data for a given value of turning speed. This can be achieved with the function
\code{turn_filt}, which can fliter angles (radians) below or above a value provided by the user. The option
exists to filter (annotate = FALSE) or flag up rows (annotate = TRUE). The function \code{adehabitatLT::as.ltraj}
is used to compute the turning angle information. \code{turn_filt} also determines if gapsections already exist
in the data and if they do, compute assessment over valid strings of points using existing gaps, or if not,
using the \code{gap_section()} defaults of GAP = 28800 s and a tol of 0.2.

Note, this function should be used with caution as behavioural characteristics of tracks may naturally be of interest ecologically,
and stripping out movements that are too straight or too variable could lead to unexpected results.
}
\examples{

indata <- yourdata # data.frame with a minimum of columns named TagID, DateTime, longitude, latitude
data <- Track(indata)
data_turn_test <- turn_filt(data, turnFilt = pi/4, turnFilt_dir = "less_than", annotate = TRUE)

# flagged angle condition
table(data_turn_test$angle_rm, useNA = "ifany") # Note, NA's flagged for start/end of a gapsection (adehabitatLT::as.ltraj needs >2) AND if any gapsection have just a single row

}
\seealso{
\link{speed_filt}
}
