% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_voltage.R
\name{Vo}
\alias{Vo}
\alias{VoCom}
\alias{Vo2}
\alias{VoCom2}
\alias{VoStack}
\alias{VoMultiStack}
\alias{VoComStack}
\alias{VoComMultiStack}
\title{Voltage ("Vo") and voltage communication ("VoCom") constructor classes for MoveRakeR}
\usage{
Vo(data)

VoCom(data)

Vo2(data)

VoCom2(data)

VoStack(data)

VoMultiStack(data)

VoComStack(data)

VoComMultiStack(data)
}
\arguments{
\item{data}{An input base::data.frame for a voltage dataset from UvA-BiTS that is to be
coerced to the designated class.}
}
\value{
For tracking data, returns a \code{Vo}, \code{VoStack} or
\code{VoMultTrackStack} from the "ee_energy_limited" UvA-BiTS database view, or
a \code{VoCom}, \code{VoComStack} or \code{VoComMultTrackStack} from the "ee_comm_limited" UvA-BiTS database view.
}
\description{
Functions to append object classes of \code{Vo}, \code{VoStack} and \code{VoMultiStack},
and \code{VoCom}, \code{VoComStack} and \code{VoComMultiStack},
obtained respectively using their respective namesake functions respectively. this is a specific
class for the UvA voltage ("Vo") and communication including voltage data ("VoCom") data.
}
\details{
Three special object classes are returned from initial the read in function
\code{\link{read_voltage_UvA}}. \code{Vo} is typically representative
of an individual animal's time-series, but an 'rbinded' \code{Vo}
object of several animals, i.e. multiple TagIDs, will also be accepted as a \code{Vo}
object and processed for each individual, thus \code{Vo} objects represent a single \code{data.frame} class;
\code{VoStack} and \code{VoMultiStack} objects are nested \code{list} formats, and (as currently implemented)
\emph{must be treated as one individual per data.frame}. These stacked listed methods
are intended for efficient application of routines for groups of animals. Objects can be coerced into
\code{Vo}, \code{VoStack} and \code{VoMultiStack} classes but they require the following
minimum column names: \strong{TagID}, \strong{DateTime} and \strong{vsll} and \strong{vbat}.
All the above is also true for the \code{VoCom} class which is here specified as a separate class
as it comes from a different part of the UvA-BiTS database with a different layout,
although it also contains voltage information, hence treated like a voltage class. The "VoCom"
class requires the column headings: \strong{TagID}, \strong{DateTime} and \strong{vbat} and \strong{bytes_communicated}.
Further checking of whether objects belong to special classes are provided by:
\code{link{is_Vo}}, \code{link{is_VoCom}},
\code{link{is_VoStack}}, \code{link{is_VoComStack}},
\code{\link{is_VoMultiStack}}, and \code{\link{is_VoComMultiStack}}and coercing of two
\code{VoStack} objects is provided by \code{\link{voltage_multi_stack}}.
}
\seealso{
\link{Track}, \link{Accn}, \link{pressure_multi_stack}, \link{is_Pr}, \link{is_PrStack}, \link{is_PrMultiStack}
}
