% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_voltage.R
\name{voltage_multi_stack}
\alias{voltage_multi_stack}
\title{voltage_multi_stack}
\usage{
voltage_multi_stack(data = NULL)
}
\arguments{
\item{data}{Object to be assesed.}
}
\value{
Return new \code{data} coerced to a \code{\link{VoMultiStack}} class.
}
\description{
Coerce \code{VoStack} class objects into a \code{VoMultiStack} class object.
}
